requirements.txt
telethon
PyGithub
requests